Code refactoring and review
Delete class : Beer, Wine, Cider, Spirit
Create class : Product, ProductSpec
Create enum : Name, Origin, Quantity, ProductType

Delete the abstract class with descendant to change by productSpec class
which going search if a spec of product searched correspond to another in
inventory. It's more flexible to do search method and to display products
in fonction to the search.

Some attributes are became enumerations. This is avoid the user's errors
when the search is enter.

Created map in productSpec to have more flexibility then the search by
enter key and value which correspond for each product.

Tested add and search methodsin findProductTest class. Errors are
displayed like it not found the spec of products.